Data not available fix for #236, #170 & maybe others
This adds a behavior for the articles_controller#show method that
redirects to the home page with an error if not all the data
that is needed is being sent to the browser. This makes for nicer
"error" messages (until we have more sophisticated JS on the
front).

Currently, the required information includes:
* @article - The article
* @commentable - a copy of the article for comments
* @comment - a new comment object
* @subjects - Associated subjects for the article

There's a helper method that mimics the logic in the
articles_controller#show method, and it's included as a way to test &
confirm that the data to be sent is valid.

- app/controllers/articles_controller.rb - Added logic to the show method
  that redirects if not all the required info is available.

- spec/controllers/articles_controller_spec.rb - Spec that confirms that
  the site will redirect if the required data above is not complete.

- app/helpers/articles_helper.rb - Added #article_sanity_check as a way
  to confirm at the view level that all required data is available for
  the template
- spec/helpers/articles_helper_spec.rb - Added success and failure specs
  for #article_sanity_check

- app/views/articles/show.html.erb - Cleaned out whitespace

- spec/views/articles/show.html.erb_spec.rb - Cleaned out whitespace