Added a javascript_lab for #116
Added a "/javascript_lab" URL that can be used by JS developers who do
not have experience with using JS within Rails to have a basic template
page useful for JS development. This page will not show up in production
but be redirected to the home page, as it is not intended to be seen
by users.

- app/controllers/static_controller.rb - Added the "javascript_lab"
  action

- spec/controllers/static_controller_spec.rb - Added success and failure
  specs for the "javascript_lab" action

- config/routes.rb - Added the mapping of "/javascript_lab" to the
  StaticController#javascript_lab action

- app/views/static/javascript_lab.html.erb - Basic template with
  information on JavaScript libraries included