Resolved problem with failing specs; #299
I fixed the problem with the three failing specs by adding a mock to the
User.all_following invoked by the mock_model in lines 5 and 27 in
`spec/mailers/user_notifier_spec.rb`. This is used by the
acts_as_followers gem. The method is mocked out to return an array with
one newly created user.