adding split gem to latest master
current initial test compares colors of follow buttons on th case show page
for ome reaon the follw button is pointing to a get request instead of
post in this context.  need to fix this issue.  otherwise gem and
dashboard appear to function correctly on local host. Need to check on
heroku with redis_to_go