Re-branched from master to add authentication for #183
also added to existing admin specs to confirm that the correct pages are
shown when the user is not logged in or is logged in as an admin user.
ultimately, we need to more gracefully handle non-admin users in this
case, but i don't think that should hold us up from pushing to
production.