for issue #365 using jquery to trucate link list
The list is currently set to show the first 10 links but we may want to
limit that further. It also includes a link to Show More to see the full
list. Still need to add a link and functioning code to Show Less as
well. The goal is to keep our article show pages from being overrun by
the long lists of links that are likely to develop as a case proceeds.