For #482 adjusting friendly_id
with this change urls will be updated when the article is updated by
setting slug: nil in the update action of the controller.

it may have been better to use a callback to set slug: nil before_save,
but this works for now. For slug uniqueness, we are set to use title,
city and zipcode within the url.