Fixing default image file
We had hard coded this to pull the file from the s3 bucket.  somehow the
file disappeared.  I added the file to the assets pipeline, but for now
we are still pulling the image from s3 but now staging.

This is not a permanent solution, but temporary until i figure out how
to get the url for the image in assets.