For #71 adding meta fields to Version
Now every version gets saved with a comment based on what is in the
article.summary field.

The next step is to make this field nil after saving. or before editing.
To avoid recursive loop it may make sense to clear the field right befor
editing instead of saving and saving again in a callback.