Minor fixes to agency creation and show
conditional limits the user to creating new agencies from the "edit"
article page. because of the redirect to lsst url, creating the agency
from the new action results in the editor being redirected to a blank
form even if they have already entered info prior to clicking to add
agency.  now the option is removed from the new form.

lastly, added a helper method to only include non-NULL fields in the
agency show view.