for #674 fixes and additions to sitemap config file
I didn't want to push this without checking in with @trystant first.
This branch contains the following changes to the sitemap config:

1. Added a line to create_index within the sitemap which causes an index
file to be created to reference multiple sitemaps

2. Added back the conditional to store sitemaps only if env==production.
Not sure why we had commented that out.

3. Added a second generator to create another sitemap with the filename
'agencies' to house urls for the agency show pages. I added the filename
so not to overwrite our current urls

In the end, though I think we want to include this new content in this
way, I decided not to push because of the small number of agencies we
have, the fact that search engines will eventually find them organically
since they are all the subject of internal links and because I don't
want to negatively affect our growing presence in search results.