Fixing most viewed list
For some reason, we were skipping the first article on the list. I
believe it related to not including the root page, but that has been
addressed.

One issue remains with thie list.  The list only prints pages that are
articles, but because other pages may get included in the query result, the
length of the list of articles may vary.

Adding scrolling to the list and having a uniform fixed length will
alleviate this issue.