for #718 add milestones and association
At a purely functional level, this works. User can add a milestone and
date_occurred to a case. User adds milestones from a prepopulated list.
Only admins can add new milestones to the list, to keep responses
uniform.

Fixes needed:

datepicker is not working on the nested form.
need to add milestones to show view in a basic timeline form.
need to add rspec for the new models