Adding scheduled task to clean backups
This task basically takes our current Autobus storage limit of 10gb and
compares that to the total size of all of our backup snapshots. If we
are over the limit, the task will delete the oldest snapshot until we
are back under the limit.

This will likely not be an issue for several months, but we are being
proactive.