for #1096 highlighting landing page metric
i took a shot at highlighting the text by increasing the size slightly, changing the background color and adding a thin outline to the text area.

I branched from the current master, so it still contains the failing test relating to the analytics page metric we have been discussing recently.