Coverage: write tests for CaseSearch
This adds tests for the `CaseSearch` class. I'm not sure why, but I
didn't need to add `search: true` to enable `Searchkick` callbacks. I
get the same test result either way, so I decided to leave it off [for
the sake of performance][perf].

I also needed to configure `Webmock` to allow connecting to `localhost`
in order to connect to `elasticsearch`. I pulled it out into its own
file even though it's only a couple of lines. `rails_helper.rb` can
become a bit of a junk drawer of configurations for a wide variety of
related gems, so I find it's nice to have each thing configured
separately.

[perf]: https://github.com/ankane/searchkick#testing