Rails upgrade 51 to 52 (#4051)
* Upgrades Rails to 5.1

Upgrades Rails to 5.1.  Fixes broken spec and bug for reverting a case

* Rubocop fixes for 5.1 upgrade

A handful of changes for Rubocop for 5.1 upgrade.  Adds new cop "Style/StringChars"

* Updates CI Postgres image

Updates the Postgres image used for CI from 11 to 12

* Adds spec for reverting new case

* Upgraded 5.1 to 5.2 in Gemfile #4037

* Added bootsnap gem #4037

* Made bootsnap require false #4037

* Added cache-version=1 value to fix github action bundle cache issue #4037

* Another approach #4037

* Removed vendored mimemagic #4037

* Added apt-get update step to resolve action build issues #4037

* Added rake to help resolve action build issue

* Reverting cache to true #4036

* Added rake gem to Gemfile.lock #4037

* Added sudo to install dependencies

* space removal #4037

* Fix CI build rubocop failure #4037

* Updated install dependencies call #4037

* Simplify update calls #4037

* Included NullFieldsCounter in DummyClass #4037

* Added lib to spec #4037

* Fixing path for lib #4037

* Manual require of lib #4037

* Updated subject factory call #4037

* Tried reverting strategy #4037

* Moved FactoryBot.use_parent_strategy line #4037

* Change build to create for factories #4037

* Fixes comment model spec for 5.2 upgrade

Fixes the comment model spec to work with the new FactoryBot syntax for the 5.2 upgrade.  Also updates the factories for cases and subjects.  Also updates the case model to remove the validation for a subject and make the cause_of_death association optional.

* Fixes spec/models/follow_spec for 5.2 upgrade

Fixes spec/models/follow_spec for the 5.2 upgrade.  Removes the follow factory as it is no longer needed.

* Added routes back into config/routes.rb #4037

* set default_url_options host #4037

* Fixes spec/models/case_spec for 5.2 upgrade

Fixes the failing specs for the followers scope for cases in the case model spec

* Fixes locale file

Restores content in locale file

* Fixes request specs for case endpoints

Fixes the request specs for case endpoints for 5.2 upgrade

* Fixes feature specs that involve displaying cases

Uses the safe navigator in case thumbnail partials to fix associated feature specs

* Fix follower test setup in specs

Fixes the way that we create followers for cases when setting up specs

* Fix null fields counter spec #4037

* Skip cause of death case rake task spec

Skips the spec for the rake task to convert cause of death data.  This spec references an association between the case model and the cause of death model that will soon be removed.

* Rubocop fixes

Co-authored-by: Mark Nyon <me@mnyon.com>
Co-authored-by: Mark Nyon <msnyon@hey.com>
Co-authored-by: Mark Nyon <mnyon@2u.com>