Updates configuration for annotate gem
The annotate gem by default places the schema information at the top of the model file as comments.  However, if there are additional comments at the beginning of the file, this can cause issues with the gem such that it redos the annotations each time and modifies the file.  This commit updates the configuration so that the schema section is placed at the bottom of the file.  It also updates the docs to include information on annotate.